+++ packages/bpk-react-utils/index.js
@@ -19,6 +19,7 @@
 import Portal from './src/Portal';
 import cssModules from './src/cssModules';
 import TransitionInitialMount from './src/TransitionInitialMount';
+import withDefaultProps from './src/withDefaultProps';
 
-export { Portal, cssModules, TransitionInitialMount };
-export default { Portal, cssModules, TransitionInitialMount };
+export { Portal, cssModules, TransitionInitialMount, withDefaultProps };
+export default { Portal, cssModules, TransitionInitialMount, withDefaultProps };