+++ packages/bpk-react-utils/src/withDefaultProps.js
@@ -19,19 +19,20 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-const withTextStyle = (WrappedComponent, textStyle, tagName, className = null) => {
+const withDefaultProps = (WrappedComponent, defaultProps) => {
+  const { className: defaultClassName, ...defaultRest } = defaultProps;
+
   const component = (props) => {
     const { children, className: innerClassName, ...rest } = props;
     const classNames = [];
 
-    if (className) { classNames.push(className); }
+    if (defaultClassName) { classNames.push(defaultClassName); }
     if (innerClassName) { classNames.push(innerClassName); }
 
     return (
       <WrappedComponent
-        textStyle={textStyle}
-        tagName={tagName}
         className={classNames.join(' ')}
+        {...defaultRest}
         {...rest}
       >
         {children}
@@ -47,9 +48,9 @@ const withTextStyle = (WrappedComponent, textStyle, tagName, className = null) =
     className: null,
   };
   const inputDisplayName = WrappedComponent.displayName || WrappedComponent.name || 'Text';
-  component.displayName = `withTextStyle(${inputDisplayName})`;
+  component.displayName = `withDefaultProps(${inputDisplayName})`;
 
   return component;
 };
 
-export default withTextStyle;
+export default withDefaultProps;